requirements.txt
Flask==3.1.0
geopy==2.4.1
Jinja2==3.1.5
meshtastic==2.5.9
pydantic==2.10.4
python-telegram-bot==20.0
